Timetable category changes.
Adding flags on the timetabl event categories.
The flags are show on dashboard and reminder settings.

If the show on dashboard is set to yes, then these events will appear on the user dashboard.
If set to no, then it will not appear here. This is useful for routine events like sleep, or events that one does not wish to accidentally be viewed on log in.

Reminder flag will b useful when the mobile component comes into existence.Some events are only there for reference or for catgories of events that do not need reminders.