Expense-types and expenses. In order to save time when capturing expenses, The ability to add a template for an expense-type has been added..
When creating an expense-type, the user is now able to also create a 
template for the format of the content.

Simply wysiwyg text, nothing fancy. it should however make routine and 
uniform expense-type capturing more efficient.

When a user is capturing an expense, when they select and expense type.
If the type has a template it will now append the template text to the 
end of the description for the expense. The user can now insert only the
text/ values which will change based on the expense.

Added feature to expense edit as well.